rename build script & intro index.html presence conditionals

+++ build.sh
@@ -21,7 +21,12 @@ STDIN
 
 DIR=$(dirname "$(realpath "$0")")
 echo " here's the directory we're in and are copying index from; $DIR"
-scp $SCP_ARGS $DIR/index.html $USER@$SERVER:~/index.html && echo "successfully copied ~/index to server"
+if [ -e $DIR/index.html ]
+  scp $SCP_ARGS $DIR/index.html $USER@$SERVER:~/index.html && echo "successfully copied ~/index to server"
+else
+ echo "This script needs $DIR/index.html to exist, which it currently does not"
+fi
+
 
 ssh $SSH_ARGS 'bash -s' <<-STDIN
   set -euo pipefail
